ng2: (common) refactor to shared polish
