+++ generators/client-2/templates/src/main/webapp/app/shared/auth/_auth.service.ts
@@ -3,7 +3,7 @@ import { StateService } from 'ui-router-ng2';
 import { LoginService } from "../login/login.service";
 import { Principal } from './principal.service';
 import { AuthServerProvider } from './auth-session.service';
-import { <%=jhiPrefixCapitalized%>TrackerService } from '../../admin/tracker/tracker.service';
+import { <%=jhiPrefixCapitalized%>TrackerService } from '../';
 
 @Injectable()
 export class AuthService {
